refactor: file_view model and improve in-place file deletion
- Split file_view/model into row_obj.rs (FileRow/FileStore) and row_imp.rs
- Rename FileView::store() to list_model(), add list_store() helper
- Move selection lookup logic (selected/selected_store) into navigation.rs
  as selected()/selected_model()
- Add FileView::remove_row() and trigger_selection_changed()
- Add ensure_selected_filter() to reselect a valid row after removal
- Update delete_current_file to remove the row directly and reselect
  instead of reloading the whole backend after permanent delete
