+++ src/file_view/navigation.rs
@@ -33,6 +33,7 @@ impl FileView {
         }
     }
 
+    /// Returns true if navigation was successful, false if there was no matching item in the given direction.
     pub fn navigate_item_bool(&self, direction: Direction, filter: &Filter, count: u32) -> bool {
         if let Some((_, new_idx)) = self.try_navigate(direction, filter, count) {
             self.select_index(new_idx);
@@ -51,12 +52,11 @@ impl FileView {
             return None;
         }
 
-        let store = self.store()?;
-        let (mut file_row, mut position) = self.selected_store(&store)?;
+        let store = self.list_model()?;
+        let (mut file_row, mut position) = self.selected_model(&store)?;
 
         let mut cnt = count;
         let n_items = store.n_items();
-
         while !filter.matches(file_row.classification()) {
             position = match direction {
                 Direction::Up => position.checked_sub(1),
@@ -104,6 +104,55 @@ impl FileView {
             }
         }
         Some(last_match)
+    }
+
+    pub fn selected(&self) -> Option<(FileRow, u32)> {
+        let model = self.list_model()?;
+        self.selected_model(&model)
+    }
+
+    pub fn selected_model(&self, model: &gio::ListModel) -> Option<(FileRow, u32)> {
+        if let Some(selection_model) = self
+            .model()
+            .and_then(|m| m.downcast::<gtk4::SingleSelection>().ok())
+        {
+            let selected_idx = selection_model.selected();
+            if selected_idx != gtk4::INVALID_LIST_POSITION {
+                if let Some(obj) = model.item(selected_idx) {
+                    if let Ok(file_row) = obj.downcast::<FileRow>() {
+                        return Some((file_row, selected_idx));
+                    }
+                }
+            }
+        }
+        if model.n_items() > 0 {
+            if let Some(obj) = model.item(0) {
+                if let Ok(file_row) = obj.downcast::<FileRow>() {
+                    return Some((file_row, 0));
+                }
+            }
+        }
+        None
+    }
+
+    pub fn ensure_selected_filter(&self, filter: &Filter) {
+        if let Some((row, _)) = self.selected() {
+            if filter.matches(row.classification()) {
+                self.trigger_selection_changed();
+            } else {
+                if self.navigate_item_bool(Direction::Down, filter, 1) {
+                    return;
+                }
+                if self.navigate_item_bool(Direction::Up, filter, 1) {
+                    return;
+                }
+                // If we reach here, there are no matching items in either direction.
+                // We show the current selection even if it doesn't match the filter,
+                // to avoid having no selection at all.
+                self.trigger_selection_changed();
+                // self.select_index(gtk4::INVALID_LIST_POSITION); // set selection to none
+            }
+        }
     }
 }
